backend/pyproject.toml
[project]
name = "retrieval-service"
version = "0.1.0"
description = "Google OAuth service with FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.8.5",
    "supabase>=2.24.0",
    "easyocr>=1.7.2",
    "openai>=2.8.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "pymupdf>=1.26.6",
    "openpyxl>=3.1.5",
    "mammoth>=1.11.0",
    "xlrd>=2.0.2",
    "rapidfuzz>=3.14.3",
    "tenacity>=9.1.2"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "hypothesis>=6.92.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["retrieval_service"]

